Removes separate *Supplier classes if not needed.
