+++ spring-cloud-gateway-server-mvc/src/main/resources/META-INF/spring.factories
@@ -1,12 +1,12 @@
 org.springframework.cloud.gateway.server.mvc.filter.FilterSupplier=\
   org.springframework.cloud.gateway.server.mvc.filter.Bucket4jFilterFunctions.FilterSupplier,\
-  org.springframework.cloud.gateway.server.mvc.filter.FilterFunctionsFilterSupplier,\
-  org.springframework.cloud.gateway.server.mvc.filter.CircuitBreakerFilterFunctionsFilterSupplier
+  org.springframework.cloud.gateway.server.mvc.filter.CircuitBreakerFilterFunctions.FilterSupplier,\
+  org.springframework.cloud.gateway.server.mvc.filter.FilterFunctions.FilterSupplier
 
 org.springframework.cloud.gateway.server.mvc.handler.HandlerSupplier=\
-  org.springframework.cloud.gateway.server.mvc.handler.HandlerFunctionsHandlerSupplier,\
+  org.springframework.cloud.gateway.server.mvc.handler.HandlerFunctions.HandlerSupplier,\
   org.springframework.cloud.gateway.server.mvc.filter.LoadBalancerHandlerSupplier
 
 org.springframework.cloud.gateway.server.mvc.predicate.PredicateSupplier=\
   org.springframework.cloud.gateway.server.mvc.predicate.MvcPredicateSupplier,\
-  org.springframework.cloud.gateway.server.mvc.predicate.GatewayPredicateSupplier
+  org.springframework.cloud.gateway.server.mvc.predicate.GatewayRequestPredicates.PredicateSupplier
